Add Ethical Book Search to Canadian, UK and US stores
Ethical Book Search provides a price comparisons between a range of book
sellers who sell both ebooks, audiobooks and print who are far more
responsible than Amazon.

Adding this to the list of book sellers makes it easier for users of
this extension to compare a range of alternatives to amazon.