Update Ethical Book Search links
To link to the search. This will be a better experience for users
because if the term is an ISBN number it will redirect to a buying
options for just that edition and if it is anything else, for example, a
title then it will present the most relevant results.